Move class component analysis into separate file
